requirements.txt
asgiref==3.5.2
Django==3.2.14
djangorestframework==3.12.4
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
psycopg2>=2.8.6,<2.9
